+++ src/37FE0.c
@@ -6,7 +6,7 @@ extern s32 func_80041FB4(s32 arg0);
 extern void func_80042034(s32 arg0);
 extern void func_800428C8(s32 arg0);
 extern void func_80071824(s32 arg0, void (*arg1)(void));
-extern void func_8003C0A4(s32 arg0, s32 arg1, s32 arg2, s32 arg3);
+extern void func_8003C0A4(s16 arg0, s16 arg1, u8 arg2, u8 arg3);
 extern void func_8003CB78(s32 arg0, s32 arg1);
 extern void func_8003C420(s32 arg0, s32 arg1, s32 arg2);
 extern void func_8003D068(s32 arg0, s32 arg1);
